Custom fields for proposals (#74)
* update gemfiles

* update mimemagic

* add codemirror for styles

* create controllers for custom fields

* add codemirror

* fix tests for codemirror editor

* order locales

* add CSS validator

* lint

* add yaml validator

* refactor propoposal presenter

* relax sassc restriction

* use formBuilder to define fields

* functional proposal creator

* fix working beta

* add asset to sprockets

* refactor markdown editor behaviour. convert to html before saving

* fix tests

* remove legacy markdown calls

* legacy markdown editor fixes

* fix legacy tests

* simplify tests

* add lib and commands tests

* add functional tests

* fix copy/paste styles in lables

* admin custom fields creation

* split tests

* refactor proposal wizard form

* process amendments. fix presentation for indirect fields (selects)

* test fixes

* fix more tess

* git 0.23 builder

* override collaborative drafts editor

* fix date field presentation

* use Nokogir to parse XML. set default text to the first textarea if xml can't be parsed

* fix 0.23 compatibility

* add tests

* fix partials

* fix collaborative draft override & tests

* fix 23 test

* Fix UTF-8 support for custom fields (#94)

Closes #83

* Fix header/paragraph bug (#96)

Closes #88

* Hotfix for empty checkboxes (#102)

* Fix checkbox empty hotfix (#106)

* Fix checkbox empty hotfix

Repair other checkbox value not being remembered correctly.
Fix attribute access error.

* Include a validation against empty checkbox groups



* fix style collection

* fix conflic solving error

* linebreak module

* i18n sanity

* renamed collect function

* normalize locales

* New Crowdin updates (#118)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Czech)

* New translations en.yml (Basque)

* New translations en.yml (Italian)

* New translations en.yml (Dutch)

* New translations en.yml (Swedish)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Czech)

* New translations en.yml (Czech)

* New translations en.yml (Japanese)

* New translations en.yml (Japanese)

* New translations en.yml (Japanese)

* New translations en.yml (Japanese)

* New translations en.yml (Japanese)

* New translations en.yml (Czech)

* New translations en.yml (Czech)

* Add translation to custom fields (#109)

* Add translation to fields

* Return key if no translation was found

* Apply bugfix from decidim/decidim#7224 (#115)


* New translations en.yml (Czech) (#121)

* update decidim test versions

* restore users_path

* add default constraint when creating a box

* remove html when assigning non xml results to bare textarea

* prevent to remove the last constraint in custom fields

* add translatable keys. add tests for not checking proposal max length

* fix admin multilang

* set form builder interface in Decidim lang

* change string literals

* fix line ending

* fix line ending

* fix eg

* bump version

* fix test conf

* fix richtext. fix & add tests

* normalize lang

* fix 0.23 compat

* New Crowdin updates (#127)

* New translations en.yml (Czech)

* New translations en.yml (Czech)

* New translations en.yml (Czech)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Basque)

* New translations en.yml (Italian)

* New translations en.yml (Japanese)

* New translations en.yml (Dutch)

* New translations en.yml (Swedish)

* New translations en.yml (Czech)

* New translations en.yml (Czech)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Basque)

* New translations en.yml (Italian)

* New translations en.yml (Japanese)

* New translations en.yml (Dutch)

* New translations en.yml (Swedish)

* New translations en.yml (Czech)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Japanese)

* New translations en.yml (Czech)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Basque)

* New translations en.yml (Italian)

* New translations en.yml (Japanese)

* New translations en.yml (Dutch)

* New translations en.yml (Swedish)

* New translations en.yml (Czech)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Basque)

* New translations en.yml (Italian)

* New translations en.yml (Japanese)

* New translations en.yml (Dutch)

* New translations en.yml (Swedish)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Czech)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Basque)

* New translations en.yml (Italian)

* New translations en.yml (Japanese)

* New translations en.yml (Dutch)

* New translations en.yml (Swedish)

* New translations en.yml (Czech)

* New translations en.yml (Czech)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* simplify function override

* fix overrides

* group editor (#129)

* introduce scope label edit (wip)

* generalize label edit

* rename fixes

* set tests

* fix locales

* fix cop and test

* fix assign empty html from xml

* New Crowdin updates (#130)

* New translations en.yml (Czech)

* New translations en.yml (French)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Basque)

* New translations en.yml (Italian)

* New translations en.yml (Japanese)

* New translations en.yml (Dutch)

* New translations en.yml (Swedish)

* New translations en.yml (Czech)

* New translations en.yml (Spanish)

* New translations en.yml (Catalan)

* New translations en.yml (Japanese)

* check precompile produciton in tests

* rollback form-builder library for sprockets compatibility

* fix previous text not being html

* (wip) fix scoped admins renames

* fix multiselect label title change

* New Crowdin updates (#131)

* New translations en.yml (Czech)

* New translations en.yml (Italian)

* fix checkboxes required fields

* fix radio buttons other val

* refactor proposal wizard form

* javascript fixes. validation fixes

+++ app/commands/decidim/decidim_awesome/admin/destroy_constraint.rb
@@ -4,6 +4,8 @@ module Decidim
   module DecidimAwesome
     module Admin
       class DestroyConstraint < Rectify::Command
+        include NeedsConstraintHelpers
+
         # Public: Initializes the command.
         #
         # constraint - A constraint constraint
@@ -18,6 +20,8 @@ module Decidim
         #
         # Returns nothing.
         def call
+          return broadcast(:invalid, I18n.t("cannot_be_destroyed", scope: "decidim.decidim_awesome.admin.config.constraints")) unless constraint_can_be_destroyed?(constraint)
+
           constraint.destroy!
           broadcast(:ok)
         rescue StandardError => e
